Creacion de carpetas public y admin

+++ server/src/controllers/products.js
@@ -32,8 +32,8 @@ export const getProduct = async (req, res) => {
 
 export const createProduct = async (req, res) => {
   try {
-    const { name, description, type, price, sku, sizes } = req.body;
-
+    const { name, description, type, price, sku, sizes, outstanding, sold } =
+      req.body;
 
     const product = new Product({
       name,
@@ -42,16 +42,20 @@ export const createProduct = async (req, res) => {
       price,
       sku,
       sizes,
+      outstanding,
+      sold,
     });
 
-    if(req.files?.image) {
-      const { public_id, secure_url } = await uploadImage( req.files.image.tempFilePath )
+    if (req.files?.image) {
+      const { public_id, secure_url } = await uploadImage(
+        req.files.image.tempFilePath
+      );
       product.image = {
         public_id,
-        secure_url
-      }
+        secure_url,
+      };
 
-      await unlink(req.files.image.tempFilePath)
+      await unlink(req.files.image.tempFilePath);
     }
 
     await product.save();
@@ -73,9 +77,10 @@ export const deleteProduct = async (req, res) => {
         message: "Producto no se puede eliminar porque no existe",
       });
 
-      await deleteImage(product.image.public_id)
-      await unlink(req.files.image.tempFilePath)
-
+    if (req.files?.image) {
+      await deleteImage(product.image.public_id);
+      await unlink(req.files.image.tempFilePath);
+    }
 
     return res.json(product);
   } catch (error) {
